fix: rename legacy src files to .cjs for ESM/CJS coexistence with v2 monorepo
