+++ package.json
@@ -10,12 +10,12 @@
     "format": "prettier --write 'packages/*/src/**/*.ts'"
   },
   "devDependencies": {
-    "typescript": "^5.7.0",
-    "vitest": "^3.1.0",
-    "tsup": "^8.4.0",
+    "@types/node": "^22.10.0",
     "eslint": "^9.16.0",
     "prettier": "^3.4.0",
-    "@types/node": "^22.10.0"
+    "tsup": "^8.4.0",
+    "typescript": "^5.7.0",
+    "vitest": "^3.1.0"
   },
   "engines": {
     "node": ">=20.0.0"
@@ -24,5 +24,13 @@
     "onlyBuiltDependencies": [
       "esbuild"
     ]
+  },
+  "dependencies": {
+    "clean-css": "^5.3.3",
+    "commander": "^15.0.0",
+    "cssom": "^0.5.0",
+    "eventproxy": "^1.0.0",
+    "pngjs": "^7.0.0",
+    "underscore": "^1.13.8"
   }
 }
