+++ src/BackgroundInterpreter.cjs
