+++ src/GrowingPacker.cjs
